MNT: move generalized PSD calculation method to the `psd` module
